Replace qVariantFromValue by QVariant::fromValue and remove some other deprecated API.

+++ src/apps/marble-ui/MarbleTest.cpp
@@ -62,7 +62,7 @@ void MarbleTest::timeDemo()
         }
     }
 
-    qDebug( "Timedemo finished in %ims", t.elapsed() );
+    qDebug( "Timedemo finished in %ims", static_cast<int>(t.elapsed()));
     qDebug() <<  QString("= %1 fps").arg(200*1000/(qreal)(t.elapsed()));
 
 }
